split the code about resource event into event_handler.go
